Switch from ZMQ to valve game sockets

+++ Code/Applications/ResourceServer/ResourceCompilationRequest.h
@@ -27,7 +27,7 @@ namespace KRG::Resource
     public:
 
         // Get the client that requested this resource
-        inline uint64 GetClientID() const { KRG_ASSERT( !IsInternalRequest() ); return m_clientID; }
+        inline uint32 GetClientID() const { KRG_ASSERT( !IsInternalRequest() ); return m_clientID; }
 
         // Get the resource ID for this request
         inline ResourceID const& GetResourceID() const { return m_resourceID; }
@@ -69,13 +69,13 @@ namespace KRG::Resource
 
     public:
 
-        uint64                              m_clientID = 0;
+        uint32                              m_clientID = 0;
         ResourceID                          m_resourceID;
         int32                               m_compilerVersion = -1;
         uint64                              m_fileTimestamp = 0;
         uint64                              m_sourceTimestampHash = 0;
-        FileSystem::Path                      m_sourceFile;
-        FileSystem::Path                      m_destinationFile;
+        FileSystem::Path                    m_sourceFile;
+        FileSystem::Path                    m_destinationFile;
         String                              m_compilerArgs;
 
         TimeStamp                           m_timeRequested;
